Extract out ideal graph
